requirements.txt
streamlit>=1.28.0
python-pptx>=0.6.21
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
openai>=1.12.0